requirements.txt
torch>=2.2
torchvision>=0.17
diffusers>=0.29.0
transformers>=4.41.0
safetensors>=0.4.2
Pillow>=10.0.0
tqdm>=4.66
scipy>=1.12
numpy>=1.26
tensorboard>=2.16
